authentication eklendi

+++ src/services/activityService.js
@@ -1,21 +1,29 @@
 import Activity from '../models/activity.js';
 
 
+const getById = async (id) => {
+  try {    
+    const activity = await Activity.findOne({ where: { id:id } });
 
+    if(!activity) throw new Error(`Activity is not found`);
+    return activity;
+  } catch (error) {
+    throw new Error(`Error in getById service: ${error.message}`);
+  }
+};
 // Function to get all activities
-const getAllActivities = async () => {
-    try {
-         const offset = (page - 1) * size;    
-        const activities = await Activity.findAll({
-          offset: offset,
-          limit: size,
-        });
+const getAllActivities = async (page, size) => {
+  try {
+    const offset = (page - 1) * size;
+    const activities = await Activity.findAll({
+      offset: offset,
+      limit: size,
+    });
 
-        
-        return activities;
-    } catch (error) {
-        throw new Error(`Error in getAllActivities service: ${error.message}`);
-    }
+    return activities;
+  } catch (error) {
+    throw new Error(`Error in getAllActivities service: ${error.message}`);
+  }
 };
 
 // Function to create a new activity
@@ -32,8 +40,12 @@ const createActivity = async (ActivityData) => {
 // Function to delete all activities
 const deleteActivity = async (activityId) => {
     try {
-        const deletedActivity = await Activity.destroy({ where: {id : +activityId },});
-        return deletedActivity;
+        const activity = await Activity.findOne({ where: { id: +activityId } });
+
+        if (!activity) throw new Error(`Activity is not found`);
+
+        await Activity.destroy({ where: {id : +activityId },});
+        return activityId;
     } catch (error) {
         throw new Error(`Error in deleteActivity service: ${error.message}`);
     }
@@ -57,4 +69,4 @@ const updateActivity = async (activityId, ActivityData) => {
 };
 
 // Export the service functions
-export default { getAllActivities, createActivity, deleteActivity, updateActivity };
+export default { getAllActivities, createActivity, deleteActivity, updateActivity,getById };
