+++ src/services/userService.js
@@ -1,4 +1,49 @@
 import User from "../models/user.js";
+import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
+
+const login = async function (email, password) {
+  try {
+    const user = await User.findOne({ email: email });
+    if (!user) {
+      throw new Error("Email or Password wrong");
+    }
+    const isValid = await bcrypt.compare(password, user.password);
+    if (!isValid) {
+      throw new Error("Email or Password wrong");
+    }
+
+    const token = await createToken(user);
+
+    return token;
+  } catch (error) {
+    console.log("Something went wrong: Service: login", error);
+    throw new Error(error);
+  }
+};
+
+const createToken = async function (user) {
+  try {
+    const token = jwt.sign({ userId: user.id }, process.env.SECRET_KEY, {
+      expiresIn: process.env.EXPIRATION_DATE,
+    });
+    return { accessToken: token};
+  } catch (error) {
+    console.log("Something went wrong: Service: login", error);
+    throw new Error(error);
+  }
+};
+
+const getById = async (id) => {
+  try {
+    const user = await User.findOne({ where: { id: id } });
+
+    if (!user) throw new Error(`User is not found`);
+    return user;
+  } catch (error) {
+    throw new Error(`Error in getById service: ${error.message}`);
+  }
+};
 
 const getAllUsers = async (page, size) => {
   try {
@@ -13,7 +58,10 @@ const getAllUsers = async (page, size) => {
 
 const createUser = async (UserData) => {
   try {
-    const newUser = await User.create(UserData);
+    const { username, email, password } = UserData;
+    const hashedpassword = await bcrypt.hash(password, 12);
+
+    const newUser = await User.create({ username, email, password:hashedpassword });
     newUser.save();
     return newUser;
   } catch (error) {
@@ -23,8 +71,11 @@ const createUser = async (UserData) => {
 
 const deleteUser = async (UserId) => {
   try {
+    const user = await User.findOne({ where: { id: +UserId } });
+
+    if (!user) throw new Error(`User is not found`);
+
     await User.destroy({ where: { id: +UserId } });
-    ("");
 
     return UserId;
   } catch (error) {
@@ -48,4 +99,4 @@ const updateUser = async (UserId, UserData) => {
   }
 };
 
-export default { getAllUsers, createUser, deleteUser, updateUser };
+export default { getAllUsers, createUser, deleteUser, updateUser, getById , login};
